requirements.txt
azure-cli==2.0.70
prometheus_client==0.7.1
git+https://github.com/rthalley/dnspython.git@f5d8cc566ededacfe4ea2cade6445e25b5fe0fe1
requests==2.21.0
